requirements.txt
Cheetah==2.4.4
Django==1.8.2
Landscape-Client==14.12
PAM==0.4.2
Pillow==2.8.2
PyYAML==3.10
Twisted-Core==13.2.0
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.3.2
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
django-blog-zinnia==0.15.2
django-contrib-comments==1.6.1
django-envelope==1.0
django-honeypot==0.5.0
django-mptt==0.7.4
django-tagging==0.4
django-tinymce==1.5.3
django-xmlrpc==0.1.5
gevent==1.0
greenlet==0.4.2
gunicorn==17.5
html5lib==0.999
jsonpatch==1.3
jsonpointer==1.0
mots-vides==2015.5.11
oauth==1.0.1
prettytable==0.7.2
psycopg2==2.4.5
pyOpenSSL==0.13
pyparsing==2.0.3
pyserial==2.6
python-apt==0.9.3.5ubuntu1
python-debian==0.1.21-nmu2ubuntu2
pytz==2015.4
requests==2.2.1
six==1.5.2
sorl-thumbnail==12.2
ssh-import-id==3.21
urllib3==1.7.1
virtualenv==1.11.4
wsgiref==0.1.2
zinnia-wysiwyg-tinymce==1.2.1
zope.interface==4.0.5
